Initial support for building TensorFlow on Windows (#17). (#4778)
* Initial support for building TensorFlow on Windows (#17).

This PR contains an initial version of support for building TensorFlow
(CPU only) on Windows using CMake. It includes documentation for building
with CMake on Windows, platform-specific code for implementing core functions
on Windows, and CMake rules for building the C++ example trainer
program and a PIP package (Python 3.5 only). The CMake rules support building
TensorFlow with Visual Studio 2015.

Windows support is a work in progress, and we welcome your feedback and
contributions.

For full details of the features currently supported and instructions for
how to build TensorFlow on Windows, please see the file
`tensorflow/contrib/cmake/README.md`.

* Exclude platform/windows where necessary.