Modifies estimator predict API to stop after a single epoch if no queue_runner.
Previously, the predict API of the Evaluator, when called with only
non-deprecated arguments, always returns a never-ending generator.
This is counterintuitive and runs against our public documentation.
This change modifies the predict API to only run through the input
examples once, if there are no queue-runners and therefore the second
epoch will be identical to the first.
Change: 140854520