GetConvolve*Algorithms fixup take 2
Move loop to toggle tensor_ops inside GetConvolveAlgorithms functions. Also
tensor_ops are not included in the returned list if they are not supported by
the cuDNN or GPU architecture versions.

This is a re-submit of PR 13252 which seems to have been accidentally squashed
during the merge at hash 935ac43.