Water down some tests so they work in the Python 2.7.9 release builds
Removing reference cycle checks from the tests that failed in
http://ci.tensorflow.org/view/Release/job/release-debian-cpu/99/consoleFull

uuid4() created reference cycles until Python 2.7.11. Removes checks for
reference cycles from tests which indirectly call it.

This issue will probably keep coming up until we move off of 2.7.9 (since
there's no presubmit), but this CL is an effort to fix the issues that came up
for the 1.7 release.

PiperOrigin-RevId: 191959519