Removed redundant tf.exp on predictions in evaluate method
In the `evaluate` method, the usage of `tf.exp` seems redundant here, since `predictions` itself is the logits output of a dense layer in decoder. Therefore, I removed `tf.exp` so that `tf.multinomial` is applied on the logits directly, which agrees better with the proper usage of `tf.multinomial`.