Remove explicit dependencies tests have on keras-related targets.
Changing all of these deps to be on a single large target. This is
because when we refactor the build files we will want only a few public
targets that things outside of Keras can depend on. And, it will make
copybara transforms much easier (can swap out single target with dummy
in opensource).

PiperOrigin-RevId: 227449197