[XLA] Add a RedzoneAllocator, and use it for checking conv correctness.
Specifically, this checks for out-of-bounds reads and writes in cudnn.

This is a BACKWARDS-INCOMPATIBLE CHANGE to autotuning.proto.  We agreed this is
necessary to clean it up.

Incidentally this cleans up some code in the use of the autotuning proto.  Code
to sort autotuning results always checked whether the run was successful, but
in fact runs were never marked as failing.  (Runs that failed were simply never
entered into the vector of results.)  So I removed the checks for success.

PiperOrigin-RevId: 240903636